# backend/openapi.yaml
openapi: 3.0.3
info:
  title: Stellar-Save Backend API
  description: |
    REST API for the Stellar-Save decentralised rotational savings platform.

    ## Authentication
    Most endpoints are public. Endpoints under `/api/user/*` require a Bearer JWT
    obtained from `POST /api/auth/verify`. Pass it as:
    ```
    Authorization: Bearer <token>
    ```

    ## Versioning
    The API is versioned via URL path prefix:
    - `/api/v1/` — stable, production-ready
    - `/api/v2/` — current development version (some routes return 501)

    Unversioned legacy paths (`/health`, `/recommendations`, etc.) are deprecated
    and will be removed on 2027-01-01.

    ## Rate Limiting
    - General endpoints: 100 req / 15 min per IP
    - Auth & admin endpoints: 10 req / 15 min per IP
    - Analytics write endpoints: stricter limits apply (see individual routes)
  version: 1.0.0
  contact:
    name: Stellar-Save
    url: https://github.com/Xoulomon/Stellar-Save
servers:
  - url: http://localhost:3001
    description: Local development
  - url: https://api.stellar-save.app
    description: Production
...
paths:
...
  /api/auth/challenge:
    post:
      tags: [Auth]
      summary: Request a sign challenge
      description: |
        Generates a one-time challenge string the client must sign with their
        Stellar Ed25519 keypair. The challenge expires in 5 minutes.
      operationId: authChallenge
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [walletAddress]
              properties:
                walletAddress:
                  type: string
                  description: Stellar public key (G... address)
                  example: GABC1234EXAMPLESTELLARADDRESS
      responses:
        '200':
          description: Challenge issued
          content:
            application/json:
              schema:
                type: object
                properties:
                  challenge:
                    type: string
                    description: One-time challenge string to sign
        '400':
          $ref: '#/components/responses/BadRequest'
  /api/auth/verify:
    post:
      tags: [Auth]
      summary: Verify signature and receive JWT
      description: |
        Verifies the Ed25519 signature against the stored challenge.
        On success, issues a signed JWT valid for 24 hours.
        The `signature` must be the base64-encoded Ed25519 signature of the
        challenge string produced by the wallet's private key.
      operationId: authVerify
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [walletAddress, challenge, signature]
              properties:
                walletAddress:
                  type: string
                  example: GABC1234EXAMPLESTELLARADDRESS
                challenge:
                  type: string
                  description: The challenge string returned by /api/auth/challenge
                signature:
                  type: string
                  description: Base64-encoded Ed25519 signature of the challenge
      responses:
        '200':
          description: Authentication successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  token:
                    type: string
                    description: Signed JWT valid for 24 hours
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'

  # ── User ──────────────────────────────────────────────────────────────────
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT issued by POST /api/auth/verify
...
  responses:
    BadRequest:
      description: Invalid request parameters
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Missing or invalid authentication
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
...